assistant_service/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
openai==1.78.0
pydantic==2.8.2
python-dotenv==1.0.1
httpx>=0.27.0
